Add ModelTrait for more reusability
